Test added, strlen : change protorype

+++ libft/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: ldevelle <ldevelle@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/12 11:30:13 by ldevelle          #+#    #+#             */
-/*   Updated: 2018/11/16 22:10:11 by ldevelle         ###   ########.fr       */
+/*   Updated: 2018/11/21 18:27:18 by ldevelle         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_strlen(const char *s)
+size_t		ft_strlen(const char *s)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (s[i])
